backend/requirements.txt
pytesseract==0.3.10
Pillow==10.2.0
pdf2image==1.16.3
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
python-magic==0.4.27 